Refactor battle logging into dedicated service
